+++ examples/intel/ip/geolocate/src/main/java/cloud/pangeacyber/examples/App.java
@@ -39,7 +39,7 @@ public class App
     			new IPGeolocateRequest.Builder(ip).provider("digitalelement").verbose(true).raw(true).build()
             );
         } catch (Exception e){
-            System.out.println("Fail to perfom request: " + e);
+            System.out.println("Request failed: " + e);
             System.exit(1);
         }