+++ examples/intel/ip/get_domain/src/main/java/cloud/pangeacyber/examples/App.java
@@ -37,7 +37,7 @@ public class App
 	    		new IPDomainRequest.Builder(ip).provider("digitalelement").verbose(true).raw(true).build()
     		);
         } catch (Exception e){
-            System.out.println("Fail to perfom request: " + e);
+            System.out.println("Request failed: " + e);
             System.exit(1);
         }