+++ examples/intel/ip/reputation_crowdstrike/src/main/java/cloud/pangeacyber/examples/App.java
@@ -38,7 +38,7 @@ public class App
                 new IPReputationRequest.Builder(ip).provider("crowdstrike").verbose(true).raw(true).build()
             );
         } catch (Exception e){
-            System.out.println("Fail to perfom request: " + e);
+            System.out.println("Request failed: " + e);
             System.exit(1);
         }