+++ examples/intel/ip/reputation_cymru/src/main/java/cloud/pangeacyber/examples/App.java
@@ -37,7 +37,7 @@ public class App
                 new IPReputationRequest.Builder(ip).provider("cymru").verbose(true).raw(true).build()
             );
         } catch (Exception e){
-            System.out.println("Fail to perfom request: " + e);
+            System.out.println("Request failed: " + e);
             System.exit(1);
         }