+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/Response.java
@@ -1,5 +1,6 @@
 package cloud.pangeacyber.pangea;
 
+import cloud.pangeacyber.pangea.exceptions.AcceptedResult;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -15,6 +16,9 @@ public class Response<Result> extends ResponseHeader {
 	CloseableHttpResponse httpResponse;
 
 	@JsonIgnore
+	AcceptedResult acceptedResult;
+
+	@JsonIgnore
 	String body = "";
 
 	public Response() {}
@@ -27,15 +31,24 @@ public class Response<Result> extends ResponseHeader {
 		return httpResponse;
 	}
 
-	protected void setHttpResponse(CloseableHttpResponse httpResponse) {
-		this.httpResponse = httpResponse;
+	protected void setHttpResponse(InternalHttpResponse httpResponse) {
+		this.httpResponse = httpResponse.getResponse();
+		this.body = httpResponse.getBody();
 	}
 
 	public String getBody() {
 		return body;
 	}
 
-	public void setBody(String body) {
-		this.body = body;
+	public AcceptedResult getAcceptedResult() {
+		return acceptedResult;
+	}
+
+	public Response(Response<?> response, AcceptedResult acceptedResult) {
+		super(response);
+		this.acceptedResult = acceptedResult;
+		this.result = null;
+		this.body = response.getBody();
+		this.httpResponse = response.getHttpResponse();
 	}
 }