+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/audit/AuditClient.java
@@ -3,6 +3,8 @@ package cloud.pangeacyber.pangea.audit;
 import cloud.pangeacyber.pangea.BaseClient;
 import cloud.pangeacyber.pangea.BaseRequest;
 import cloud.pangeacyber.pangea.Config;
+import cloud.pangeacyber.pangea.PostConfig;
+import cloud.pangeacyber.pangea.Response;
 import cloud.pangeacyber.pangea.audit.arweave.*;
 import cloud.pangeacyber.pangea.audit.models.*;
 import cloud.pangeacyber.pangea.audit.requests.*;
@@ -17,6 +19,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -34,14 +37,42 @@ final class RootRequest extends BaseRequest {
 	}
 }
 
-final class LogRequest extends BaseRequest {
+final class LogEvent {
 
 	@JsonProperty("event")
 	IEvent event;
 
 	@JsonInclude(Include.NON_NULL)
-	@JsonProperty("verbose")
-	Boolean verbose;
+	@JsonProperty("signature")
+	String signature;
+
+	@JsonInclude(Include.NON_NULL)
+	@JsonProperty("public_key")
+	String publicKey;
+
+	public LogEvent(IEvent event, String signature, String publicKey) {
+		this.event = event;
+		this.signature = signature;
+		this.publicKey = publicKey;
+	}
+
+	public IEvent getEvent() {
+		return event;
+	}
+
+	public String getSignature() {
+		return signature;
+	}
+
+	public String getPublicKey() {
+		return publicKey;
+	}
+}
+
+final class LogRequest extends LogCommonRequest {
+
+	@JsonProperty("event")
+	IEvent event;
 
 	@JsonInclude(Include.NON_NULL)
 	@JsonProperty("signature")
@@ -55,12 +86,23 @@ final class LogRequest extends BaseRequest {
 	@JsonProperty("prev_root")
 	String prevRoot;
 
-	public LogRequest(IEvent event, Boolean verbose, String signature, String publicKey, String prevRoot) {
-		this.event = event;
-		this.verbose = verbose;
-		this.signature = signature;
-		this.publicKey = publicKey;
+	public LogRequest(LogEvent event, Boolean verbose, String prevRoot) {
+		super(verbose);
+		this.event = event.getEvent();
+		this.signature = event.getSignature();
+		this.publicKey = event.getPublicKey();
 		this.prevRoot = prevRoot;
+	}
+}
+
+final class LogBulkRequest extends LogCommonRequest {
+
+	@JsonProperty("events")
+	ArrayList<LogEvent> events;
+
+	public LogBulkRequest(ArrayList<LogEvent> events, Boolean verbose) {
+		super(verbose);
+		this.events = events;
 	}
 }
 
@@ -139,18 +181,74 @@ public class AuditClient extends BaseClient {
 		}
 	}
 
-	private LogResponse logPost(IEvent event, Boolean verbose, String signature, String publicKey, boolean verify)
-		throws PangeaException, PangeaAPIException {
+	private LogRequest getLogRequest(LogEvent event, Boolean verbose, boolean verify) {
 		String prevRoot = null;
 		if (verify) {
 			verbose = true;
 			prevRoot = this.prevUnpublishedRoot;
 		}
-		LogRequest request = new LogRequest(event, verbose, signature, publicKey, prevRoot);
-		return post("/v1/log", request, LogResponse.class);
+		return new LogRequest(event, verbose, prevRoot);
 	}
 
 	private LogResponse doLog(IEvent event, LogConfig config) throws PangeaException, PangeaAPIException {
+		LogEvent logEvent = getLogEvent(event, config);
+		LogResponse response = post(
+			"/v1/log",
+			getLogRequest(logEvent, config.getVerbose(), config.getVerify()),
+			LogResponse.class
+		);
+		processLogResult(response.getResult(), config.getVerify());
+		return response;
+	}
+
+	private LogBulkResponse doLogBulk(IEvent[] events, LogConfig config) throws PangeaException, PangeaAPIException {
+		LogBulkResponse response = post(
+			"/v2/log",
+			new LogBulkRequest(getLogEvents(events, config), config.getVerbose()),
+			LogBulkResponse.class
+		);
+
+		if (response.getResult() != null) {
+			for (LogResult result : response.getResult().getResults()) {
+				processLogResult(result, config.getVerify());
+			}
+		}
+		return response;
+	}
+
+	private LogBulkResponse doLogBulkAsync(IEvent[] events, LogConfig config)
+		throws PangeaException, PangeaAPIException {
+		LogBulkResponse response;
+		try {
+			response =
+				post(
+					"/v2/log_async",
+					new LogBulkRequest(getLogEvents(events, config), config.getVerbose()),
+					LogBulkResponse.class,
+					new PostConfig.Builder().pollResult(false).build()
+				);
+		} catch (AcceptedRequestException e) {
+			return new LogBulkResponse(new Response<LogBulkResult>(e.getResponse(), e.getAcceptedResult()));
+		}
+
+		if (response.getResult() != null) {
+			for (LogResult result : response.getResult().getResults()) {
+				processLogResult(result, false);
+			}
+		}
+		return response;
+	}
+
+	private ArrayList<LogEvent> getLogEvents(IEvent[] events, LogConfig config) throws PangeaException {
+		ArrayList<LogEvent> logEvents = new ArrayList<LogEvent>();
+		for (IEvent event : events) {
+			LogEvent logEvent = getLogEvent(event, config);
+			logEvents.add(logEvent);
+		}
+		return logEvents;
+	}
+
+	private LogEvent getLogEvent(IEvent event, LogConfig config) throws PangeaException {
 		String signature = null;
 		String publicKey = null;
 
@@ -173,9 +271,7 @@ public class AuditClient extends BaseClient {
 			}
 		}
 
-		LogResponse response = logPost(event, config.getVerbose(), signature, publicKey, config.getVerify());
-		processLogResponse(response.getResult(), config.getVerify());
-		return response;
+		return new LogEvent(event, signature, publicKey);
 	}
 
 	private String getPublicKeyData() throws PangeaException {
@@ -197,7 +293,7 @@ public class AuditClient extends BaseClient {
 		}
 	}
 
-	private void processLogResponse(LogResult result, boolean verify) throws VerificationFailed, PangeaException {
+	private void processLogResult(LogResult result, boolean verify) throws VerificationFailed, PangeaException {
 		String newUnpublishedRoot = result.getUnpublishedRoot();
 		result.setEventEnvelope(EventEnvelope.fromRaw(result.getRawEnvelope(), (Class<IEvent>) this.customSchemaClass));
 		if (verify) {
@@ -249,6 +345,64 @@ public class AuditClient extends BaseClient {
 			config = new LogConfig.Builder().build();
 		}
 		return doLog(event, config);
+	}
+
+	/**
+	 * Log multiple entries
+	 * @pangea.description Create multiple log entries in the Secure Audit Log.
+	 * @pangea.operationId audit_post_v2_log
+	 * @param events events to log
+	 * @param config
+	 * @return LogBulkResponse
+	 * @throws PangeaException
+	 * @throws PangeaAPIException
+	 * @pangea.code
+	 * {@code
+	 * StandardEvent event = new StandardEvent
+	 * 	.Builder("Hello, World!")
+	 * 	.action("Login")
+	 * 	.actor("Terminal")
+	 * 	.build();
+	 * StandardEvent[] events = {event};
+	 * LogConfig config = new LogConfig.Builder().build();
+	 *
+	 * LogBulkResponse response = client.logBulk(events, config);
+	 * }
+	 */
+	public LogBulkResponse logBulk(IEvent[] events, LogConfig config) throws PangeaException, PangeaAPIException {
+		if (config == null) {
+			config = new LogConfig.Builder().build();
+		}
+		return doLogBulk(events, config);
+	}
+
+	/**
+	 * Log multiple entries asynchronously
+	 * @pangea.description Asynchronously create multiple log entries in the Secure Audit Log.
+	 * @pangea.operationId audit_post_v2_log_async
+	 * @param events events to log
+	 * @param config
+	 * @return LogBulkResponse
+	 * @throws PangeaException
+	 * @throws PangeaAPIException
+	 * @pangea.code
+	 * {@code
+	 * StandardEvent event = new StandardEvent
+	 * 	.Builder("Hello, World!")
+	 * 	.action("Login")
+	 * 	.actor("Terminal")
+	 * 	.build();
+	 * StandardEvent[] events = {event};
+	 * LogConfig config = new LogConfig.Builder().build();
+	 *
+	 * LogBulkResponse response = client.logBulkAsync(events, config);
+	 * }
+	 */
+	public LogBulkResponse logBulkAsync(IEvent[] events, LogConfig config) throws PangeaException, PangeaAPIException {
+		if (config == null) {
+			config = new LogConfig.Builder().build();
+		}
+		return doLogBulkAsync(events, config);
 	}
 
 	private RootResponse rootPost(Integer treeSize) throws PangeaException, PangeaAPIException {