+++ packages/pangea-sdk/src/test/java/cloud/pangeacyber/pangea/redact/ITRedactTest.java
@@ -278,4 +278,56 @@ public class ITRedactTest {
 			new RedactStructuredRequest.Builder(data).build()
 		);
 	}
+
+	@Test
+	public void testMultiConfig1Redact() throws PangeaException, PangeaAPIException, ConfigException {
+		Config cfg = new Config.Builder(Config.getMultiConfigTestToken(environment), Config.getTestDomain(environment))
+			.build();
+
+		String configID = Config.getConfigID(environment, "redact", 1);
+
+		RedactClient clientMultiConfig = new RedactClient.Builder(cfg).withConfigID(configID).build();
+
+		RedactTextResponse response = clientMultiConfig.redactText(
+			new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").build()
+		);
+		assertTrue(response.isOk());
+
+		RedactTextResult result = response.getResult();
+		assertEquals("<PERSON>... <PHONE_NUMBER>", result.getRedactedText());
+		assertEquals(2, result.getCount());
+		assertNull(result.getReport());
+	}
+
+	@Test
+	public void testMultiConfig2Redact() throws PangeaException, PangeaAPIException, ConfigException {
+		Config cfg = new Config.Builder(Config.getMultiConfigTestToken(environment), Config.getTestDomain(environment))
+			.build();
+
+		String configID = Config.getConfigID(environment, "redact", 2);
+
+		RedactClient clientMultiConfig = new RedactClient.Builder(cfg).withConfigID(configID).build();
+
+		RedactTextResponse response = clientMultiConfig.redactText(
+			new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").build()
+		);
+		assertTrue(response.isOk());
+
+		RedactTextResult result = response.getResult();
+		assertEquals("<PERSON>... <PHONE_NUMBER>", result.getRedactedText());
+		assertEquals(2, result.getCount());
+		assertNull(result.getReport());
+	}
+
+	@Test(expected = PangeaAPIException.class)
+	public void testMultiConfigWithoutConfigID() throws PangeaException, PangeaAPIException, ConfigException {
+		Config cfg = new Config.Builder(Config.getMultiConfigTestToken(environment), Config.getTestDomain(environment))
+			.build();
+
+		RedactClient clientMultiConfig = new RedactClient.Builder(cfg).build();
+
+		RedactTextResponse response = clientMultiConfig.redactText(
+			new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").build()
+		);
+	}
 }